+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: jayache <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/07 12:44:53 by jayache           #+#    #+#             */
-/*   Updated: 2018/12/22 15:53:27 by jayache          ###   ########.fr       */
+/*   Updated: 2018/12/30 17:41:30 by jayache          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 int	ft_atoi(const char *str)
 {
-	unsigned long	nb;
-	int				x;
-	int				sign;
+	uintmax_t	nb;
+	int			x;
+	int			sign;
 
 	sign = 1;
 	x = 0;
@@ -33,10 +33,10 @@ int	ft_atoi(const char *str)
 	while (str[x] >= '0' && str[x] <= '9')
 	{
 		nb *= 10;
-		nb += (str[x] - '0');
+		nb += (uintmax_t)(str[x] - '0');
 		++x;
 		if (nb > 9223372036854775807)
 			return (sign < 0 ? -0 : -1);
 	}
-	return ((int)(nb * sign));
+	return ((int)((int)nb * sign));
 }
