+++ ft_stradd.c
@@ -6,7 +6,7 @@
 /*   By: jayache <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/09 14:55:36 by jayache           #+#    #+#             */
-/*   Updated: 2018/12/10 11:05:11 by jayache          ###   ########.fr       */
+/*   Updated: 2018/12/30 16:49:47 by jayache          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_stradd(char **str, char c)
 {
-	int		s;
+	size_t	s;
 	char	*result;
 	char	*sc;
