+++ ft_vector3_magnitude.c
@@ -6,13 +6,13 @@
 /*   By: jayache <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/01 14:55:28 by jayache           #+#    #+#             */
-/*   Updated: 2018/12/01 15:19:45 by jayache          ###   ########.fr       */
+/*   Updated: 2018/12/30 17:09:29 by jayache          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-float	ft_vector3_magnitude(t_vector3 vec)
+double	ft_vector3_magnitude(t_vector3 vec)
 {
 	return (sqrt((vec.x * vec.x) + (vec.y * vec.y) + (vec.z * vec.z)));
 }
