+++ ft_vector4.c
@@ -6,13 +6,13 @@
 /*   By: jayache <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/05 11:09:51 by jayache           #+#    #+#             */
-/*   Updated: 2018/12/05 11:10:38 by jayache          ###   ########.fr       */
+/*   Updated: 2018/12/30 17:10:40 by jayache          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-t_vector4	ft_vector4(float x, float y, float z, float w)
+t_vector4	ft_vector4(double x, double y, double z, double w)
 {
 	t_vector4 vec;
